+++ js/core/src/node.ts
@@ -15,7 +15,7 @@
  */
 
 import { initNodeAsyncContext } from './node-async-context.js';
-import { initNodeTelemetry } from './tracing/node.js';
+import { initNodeTelemetry } from './tracing/node-telemetry-provider.js';
 
 export function initNodeFeatures() {
   initNodeAsyncContext();
